TCIC2-3 Documented and reorganized the admin project
