+++ _scripts/rustfmt.sh
@@ -2,4 +2,5 @@
 
 rustfmt --edition 2021 */src/*.rs
 rustfmt --edition 2021 */src/**/*.rs
+rustfmt --edition 2021 */src/**/**/*.rs
 # add more /**/ as necessary
